requirements.txt
apturl==0.5.2
asn1crypto==0.24.0
astroid==2.3.0
Brlapi==0.6.6
certifi==2018.1.18
chardet==3.0.4
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
defer==1.0.6
distro-info===0.18ubuntu0.18.04.1
dj-database-url==0.5.0
Django==2.0
django-grappelli==2.13.1
django-heroku==0.3.1
django-webpack-loader==0.6.0
djangorestframework==3.10.3
Faker==2.0.2
httplib2==0.9.2
idna==2.6
isort==4.3.21
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.4.2
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
netifaces==0.10.4
oauth==1.0.1
olefile==0.45.1
pexpect==4.2.1
Pillow==5.1.0
pipenv==2018.11.26
protobuf==3.0.0
psycopg2==2.8.4
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
pylint==2.4.3
pymacaroons==0.13.0
PyMySQL==0.9.3
PyNaCl==1.1.2
pyRFC3339==1.0
python-apt==1.6.4
python-dateutil==2.6.1
python-debian==0.1.32
pytz==2018.3
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
SecretStorage==2.3.1
simplejson==3.13.2
six==1.12.0
sqlparse==0.3.0
system-service==0.3
systemd-python==234
test==2.3.4.5
text-unidecode==1.3
typed-ast==1.2.0
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.22
usb-creator==0.3.3
virtualenv==16.7.5
virtualenv-clone==0.5.3
wadllib==1.3.2
whitenoise==4.1.4
wrapt==1.11.2
xkit==0.0.0
zope.interface==4.3.2
